requirements.txt
Flask==2.3.3
requests==2.31.0
pycryptodome==3.19.0
protobuf==3.20.3
gunicorn==21.2.0